# beryl-core-api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Beryl Core API",
    "description": "Central API Gateway and orchestration layer for Béryl ecosystem",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/esg/pedometer/data": {
      "post": {
        "tags": [
          "ESG"
        ],
        "summary": "Get Pedometer Data",
        "description": "Retrieve pedometer data for a user.\n\n- **user_id**: User identifier\n- **date_from**: Start date for data retrieval\n- **date_to**: End date for data retrieval",
        "operationId": "get_pedometer_data_api_v1_esg_pedometer_data_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PedometerRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PedometerResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
...
    "/api/v1/esg/esg/score": {
      "post": {
        "tags": [
          "ESG"
        ],
        "summary": "Calculate Esg Score",
        "description": "Calculate ESG score for a user.\n\n- **user_id**: User identifier\n- **period**: Calculation period (daily, weekly, monthly, yearly)",
        "operationId": "calculate_esg_score_api_v1_esg_esg_score_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EsgScoreRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EsgScoreResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/esg/sustainability/indicators": {
      "post": {
        "tags": [
          "ESG"
        ],
        "summary": "Get Sustainability Indicators",
        "description": "Get sustainability indicators for a user.\n\n- **user_id**: User identifier",
        "operationId": "get_sustainability_indicators_api_v1_esg_sustainability_indicators_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SustainabilityIndicatorsRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SustainabilityIndicatorsResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/esg/community/{community_id}/impact": {
      "post": {
        "tags": [
          "ESG"
        ],
        "summary": "Get Community Impact",
        "description": "Get community impact metrics.\n\n- **community_id**: Community identifier (path parameter)",
        "operationId": "get_community_impact_api_v1_esg_community__community_id__impact_post",
        "parameters": [
          {
            "name": "community_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Community Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CommunityImpactRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommunityImpactResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/esg/challenges/{user_id}/active": {
      "get": {
        "tags": [
          "ESG"
        ],
        "summary": "Get Active Challenges",
        "description": "Get active health challenges for a user.\n\n- **user_id**: User identifier (path parameter)",
        "operationId": "get_active_challenges_api_v1_esg_challenges__user_id__active_get",
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "User Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/HealthChallengeResponse"
                  },
                  "title": "Response Get Active Challenges Api V1 Esg Challenges  User Id  Active Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/esg/wellbeing/score": {
      "post": {
        "tags": [
          "ESG"
        ],
        "summary": "Get Wellbeing Score",
        "description": "Get comprehensive wellbeing score for a user.\n\n- **user_id**: User identifier",
        "operationId": "get_wellbeing_score_api_v1_esg_wellbeing_score_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WellbeingScoreRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WellbeingScoreResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/esg/reporting/esg-report": {
      "post": {
        "tags": [
          "ESG"
        ],
        "summary": "Generate Esg Report",
        "description": "Generate ESG report for institutional reporting.\n\n- **user_id**: User identifier\n- **report_type**: Report type (standard, detailed, executive)",
        "operationId": "generate_esg_report_api_v1_esg_reporting_esg_report_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EsgReportRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EsgReportResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/esg/insights/{user_id}/personalized": {
      "get": {
        "tags": [
          "ESG"
        ],
        "summary": "Get Personalized Insights",
        "description": "Get personalized ESG and health insights.\n\n- **user_id**: User identifier (path parameter)\n\nAggregates pedometer, health, ESG, and wellbeing data for comprehensive view.",
        "operationId": "get_personalized_insights_api_v1_esg_insights__user_id__personalized_get",
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "User Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PersonalizedInsightsResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CommunityImpactRequest": {
        "properties": {
          "community_id": {
            "type": "string",
            "title": "Community Id",
            "description": "Community identifier"
          }
        },
        "type": "object",
        "required": [
          "community_id"
        ],
        "title": "CommunityImpactRequest",
        "description": "Request model for community impact metrics."
      },
...
      "EsgReportRequest": {
        "properties": {
          "user_id": {
            "type": "string",
            "title": "User Id",
            "description": "User identifier"
          },
          "report_type": {
            "type": "string",
            "title": "Report Type",
            "description": "standard, detailed, executive",
            "default": "standard"
          }
        },
        "type": "object",
        "required": [
          "user_id"
        ],
        "title": "EsgReportRequest",
        "description": "Request model for ESG report generation."
      },
...
      "EsgScoreRequest": {
        "properties": {
          "user_id": {
            "type": "string",
            "title": "User Id",
            "description": "User identifier"
          },
          "period": {
            "type": "string",
            "title": "Period",
            "description": "daily, weekly, monthly, yearly",
            "default": "monthly"
          }
        },
        "type": "object",
        "required": [
          "user_id"
        ],
        "title": "EsgScoreRequest",
        "description": "Request model for ESG score calculation."
      },
      "EsgScoreResponse": {
        "properties": {
          "user_id": {
            "type": "string",
            "title": "User Id"
          },
          "environmental_score": {
            "type": "number",
            "maximum": 100.0,
            "minimum": 0.0,
            "title": "Environmental Score"
          },
          "social_score": {
            "type": "number",
            "maximum": 100.0,
            "minimum": 0.0,
            "title": "Social Score"
          },
          "governance_score": {
            "type": "number",
            "maximum": 100.0,
            "minimum": 0.0,
            "title": "Governance Score"
          },
          "overall_score": {
            "type": "number",
            "maximum": 100.0,
            "minimum": 0.0,
            "title": "Overall Score"
          },
          "period": {
            "type": "string",
            "title": "Period"
          },
          "trend": {
            "type": "string",
            "title": "Trend",
            "description": "up, stable, down"
          },
          "timestamp": {
            "type": "string",
            "format": "date-time",
            "title": "Timestamp"
          }
        },
        "type": "object",
        "required": [
          "user_id",
          "environmental_score",
          "social_score",
          "governance_score",
          "overall_score",
          "period",
          "trend",
          "timestamp"
        ],
        "title": "EsgScoreResponse",
        "description": "Response model for ESG score."
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "PedometerRequest": {
        "properties": {
          "user_id": {
            "type": "string",
            "title": "User Id",
            "description": "User identifier"
          },
          "date_from": {
            "type": "string",
            "format": "date",
            "title": "Date From",
            "description": "Start date for data retrieval"
          },
          "date_to": {
            "type": "string",
            "format": "date",
            "title": "Date To",
            "description": "End date for data retrieval"
          }
        },
        "type": "object",
        "required": [
          "user_id",
          "date_from",
          "date_to"
        ],
        "title": "PedometerRequest",
        "description": "Request model for pedometer data retrieval."
      },
      "PedometerResponse": {
        "properties": {
          "user_id": {
            "type": "string",
            "title": "User Id"
          },
          "date": {
            "type": "string",
            "format": "date",
            "title": "Date"
          },
          "steps": {
            "type": "integer",
            "title": "Steps"
          },
          "distance_km": {
            "type": "number",
            "title": "Distance Km"
          },
          "calories_burned": {
            "type": "number",
            "title": "Calories Burned"
          },
          "active_minutes": {
            "type": "integer",
            "title": "Active Minutes"
          },
          "heart_rate_avg": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Heart Rate Avg"
          },
          "timestamp": {
            "type": "string",
            "format": "date-time",
            "title": "Timestamp"
          }
        },
        "type": "object",
        "required": [
          "user_id",
          "date",
          "steps",
          "distance_km",
          "calories_burned",
          "active_minutes",
          "heart_rate_avg",
          "timestamp"
        ],
        "title": "PedometerResponse",
        "description": "Response model for pedometer data."
      },
...
      "SustainabilityIndicatorsRequest": {
        "properties": {
          "user_id": {
            "type": "string",
            "title": "User Id",
            "description": "User identifier"
          }
        },
        "type": "object",
        "required": [
          "user_id"
        ],
        "title": "SustainabilityIndicatorsRequest",
        "description": "Request model for sustainability indicators."
      },
...
      "WellbeingScoreRequest": {
        "properties": {
          "user_id": {
            "type": "string",
            "title": "User Id",
            "description": "User identifier"
          }
        },
        "type": "object",
        "required": [
          "user_id"
        ],
        "title": "WellbeingScoreRequest",
        "description": "Request model for wellbeing score."
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}